[Refactor]: Refactored and restructured packages
