+++ pkg/domain/mail/email_dto.go
@@ -1,4 +1,4 @@
-package dto
+package mail
 
 type EmailDto struct {
 	Email string `form:"email" binding:"required,email"`
