+++ pkg/domain/mail/email_send_dto.go
@@ -1,4 +1,4 @@
-package dto
+package mail
 
 type EmailSendDto struct {
 	Email       string
