+++ pkg/domain/rate/service.go
@@ -1,4 +1,4 @@
-package service
+package rate
 
 import (
 	"errors"
