+++ pkg/domain/user/service.go
@@ -1,4 +1,4 @@
-package service
+package user
 
 import (
 	"errors"
