+++ pkg/model/rate_model.go
